Fix bug with hidden login button
The same login button is used for all "authentication steps".
We need to hide that button on the first one, because login button from "PluggableAuth" is used there.
But as a result, button was hidden on the all steps. So users could not proceed, because there was no "login" button on the page.

[4.2.x]

ERM27705

Change-Id: I217142fdf1ff801f8546ac3aaf5344d83778b1cb